Refactor to ArrowFunction
